moved styles to seperate file
